perf(executor): improving Sentencizer by a factor of 2.5
The Sententicizer is now 2.5 times faster.
Instead of replacing separators by a '\n' character and splitting later
by the new line character, this improved version does that with one regex
and using the `findall` method.
The complexity of the new algorithm is also linear but with a low
constant time because it uses only one pass over the input string.
The regex also ignores spaces at the beginning and end of the sentences
so no `strip` needed.